Switch to using the strip_attributes gem
Previously, we were using a customized version of strip_attributes
in the vendor/plugins version.  Plugins in vendor/plugins are
deprecated in Rails 3.2, and removed completely in Rails 4, so it's
a good idea to switch to the gem anyway.  In addition, one of the
customizations to the plugin has now been added as an option into
the upstream, thanks to a pull request from Henare Degan:

   https://github.com/rmm5t/strip_attributes/pull/6