Fix quietly_try_to_purge on Ruby 2.0
The tests of quietly_try_to_purge were failing on Ruby 2.0 due to the
net-http-local gem not working with that Ruby version.  When
Net::HTTP::bind is called, it temporarily replaces the open method of
TCPSocket with a version of open that only takes two parameters. The
Ruby 1.9 version of net/http.rb calls TCPSocket.open with two
parameters, but the Ruby 2.0 version calls it with 4, and so fails with
a mismatched number of arguments error.

In fact, net-http-local doesn't seem to be necessary with Ruby 2.0,
where one can supply a :local_port argument to Net::HTTP.start, so this
commit patches lib/quiet_opener.rb to use that approach with Ruby
>= 2.0, and net-http-local on earlier versions.